# packages/sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders/preview": {
      "post": {
        "operationId": "OrdersController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientOrderId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64,
                    "pattern": "^[A-Za-z0-9._:-]+$"
                  },
                  "symbol": {
                    "type": "string",
                    "pattern": "^[A-Z0-9][A-Z0-9._-]{0,31}$"
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "buy",
                      "sell"
                    ]
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "market",
                      "limit",
                      "stop",
                      "stop_limit",
                      "trailing",
                      "bracket",
                      "oco"
                    ]
                  },
                  "qty": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "limitPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "stopPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "trailAmount": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "entryType": {
                    "type": "string",
                    "enum": [
                      "market",
                      "limit"
                    ]
                  },
                  "stopLossPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "takeProfitPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "legs": {
                    "minItems": 2,
                    "maxItems": 2,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "limit",
                            "stop",
                            "stop_limit"
                          ]
                        },
                        "limitPrice": {
                          "type": "string",
                          "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                        },
                        "stopPrice": {
                          "type": "string",
                          "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                        }
                      },
                      "required": [
                        "type"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "tif": {
                    "default": "gtc",
                    "type": "string",
                    "enum": [
                      "day",
                      "gtc",
                      "ioc",
                      "fok",
                      "gtd"
                    ]
                  },
                  "expireAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                  },
                  "reduceOnly": {
                    "default": false,
                    "type": "boolean"
                  },
                  "postOnly": {
                    "default": false,
                    "type": "boolean"
                  },
                  "source": {
                    "default": "manual",
                    "type": "string",
                    "enum": [
                      "manual",
                      "ai-draft-accepted"
                    ]
                  }
                },
                "required": [
                  "symbol",
                  "side",
                  "type",
                  "qty"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Order preview: notional, fees + spread, FX conversion, margin impact, loss if the stop is hit (currency and % of equity), reward:risk, confirmation flag and the pre-trade risk verdict. PAPER, SIMULATED.",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders": {
      "post": {
        "operationId": "OrdersController_place",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientOrderId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64,
                    "pattern": "^[A-Za-z0-9._:-]+$"
                  },
                  "symbol": {
                    "type": "string",
                    "pattern": "^[A-Z0-9][A-Z0-9._-]{0,31}$"
                  },
                  "side": {
                    "type": "string",
                    "enum": [
                      "buy",
                      "sell"
                    ]
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "market",
                      "limit",
                      "stop",
                      "stop_limit",
                      "trailing",
                      "bracket",
                      "oco"
                    ]
                  },
                  "qty": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "limitPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "stopPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "trailAmount": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "entryType": {
                    "type": "string",
                    "enum": [
                      "market",
                      "limit"
                    ]
                  },
                  "stopLossPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "takeProfitPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "legs": {
                    "minItems": 2,
                    "maxItems": 2,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "limit",
                            "stop",
                            "stop_limit"
                          ]
                        },
                        "limitPrice": {
                          "type": "string",
                          "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                        },
                        "stopPrice": {
                          "type": "string",
                          "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                        }
                      },
                      "required": [
                        "type"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "tif": {
                    "default": "gtc",
                    "type": "string",
                    "enum": [
                      "day",
                      "gtc",
                      "ioc",
                      "fok",
                      "gtd"
                    ]
                  },
                  "expireAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                  },
                  "reduceOnly": {
                    "default": false,
                    "type": "boolean"
                  },
                  "postOnly": {
                    "default": false,
                    "type": "boolean"
                  },
                  "source": {
                    "default": "manual",
                    "type": "string",
                    "enum": [
                      "manual",
                      "ai-draft-accepted"
                    ]
                  }
                },
                "required": [
                  "clientOrderId",
                  "symbol",
                  "side",
                  "type",
                  "qty"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Place a paper order. Idempotent on clientOrderId (201 new, 200 replay, 409 reused with a different body). Risk rejections return 422 with a machine code and a plain-language message.",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "symbol",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "open",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List orders (open by default), newest first.",
        "tags": [
          "orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_cancelAll",
        "parameters": [
          {
            "name": "symbol",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cancel every open order, optionally for one symbol (Pro terminal \"Cancel all\"; each cancel is audited).",
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "One order with its children (bracket stop/target, OCO legs).",
        "tags": [
          "orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_amend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "qty": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "limitPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "stopPrice": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  },
                  "trailAmount": {
                    "type": "string",
                    "pattern": "^[+-]?(?:\\d+)(?:\\.\\d+)?$"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Amend quantity, limit, stop or trailing distance of a working order (audited).",
        "tags": [
          "orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cancel an open order (audited). Cancelling an OCO group cancels its legs.",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "KORA API",
    "description": "KORA platform API (PAPER only). Money and prices are decimal strings.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kora_at"
      }
    },
...
  }
}